requirements.txt
APScheduler==3.9.1
attrs==22.1.0
Automat==20.2.0
Babel==2.9.1
blinker==1.5
certifi==2021.10.8
chardet==4.0.0
click==7.1.2
constantly==15.1.0
coverage==5.5
decorator==5.1.1
defusedxml==0.7.1
Django==2.2.28
dnspython==1.16.0
dominate==2.7.0
email-validator==1.3.0
Flask==1.1.4
Flask-Babel==2.0.0
Flask-Bootstrap==3.3.7.1
Flask-HTTPAuth==4.7.0
Flask-Log-Request-ID==0.10.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-OpenID==1.3.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
Flask-WhooshAlchemy==0.56
Flask-WTF==0.14.3
flipflop==1.0
greenlet==1.1.3.post0
guess-language==0.2
gunicorn==20.1.0
httpauth==0.3
hyperlink==21.0.0
idna==2.10
incremental==22.10.0
itsdangerous==1.1.0
Jinja2==2.11.3
lxml==4.9.1
MarkupSafe==1.1.1
pbr==5.10.0
psycopg2-binary==2.8.6
PyJWT==1.7.1
#python-docx==0.8.11
python3-openid==3.2.0
pytz==2022.4
PyYAML==5.3.1
redis==3.5.3
requests==2.25.1
six==1.16.0
sortedcontainers==2.4.0
SQLAlchemy==1.3.24
sqlalchemy-migrate==0.13.0
sqlparse==0.4.3
Tempita==0.5.2
tornado==6.1
tzlocal==2.1
ua-parser==0.16.1
urllib3==1.26.9
user-agents==2.2.0
visitor==0.1.3
Werkzeug==1.0.1
Whoosh==2.7.4
wtf==0.1
WTForms==2.3.3
xlwt==1.3.0
